Minor improvements
- Aligned InBetween with dotnet naming of IsBetween
- Simplified MoveList
- Spelling fix in HiResTimer
- SearchParameter depth now int + added missing method to interface
- Encapsulated UCI options for simpler use
- Added bitboard debugger display
- Some types now implement IMinMaxValue<T>
- Added some static square properties with tests
- Minor formatting

+++ src/Rudzoft.ChessLib/Protocol/UCI/SearchParameters.cs
@@ -98,7 +98,7 @@ public sealed class SearchParameters : ISearchParameters
         set => _movesToGo = value;
     }
 
-    public ulong Depth { get; set; }
+    public int Depth { get; set; }
 
     public ulong Nodes { get; set; }
